Update README.md
Modernizing how people interact with gold investments, Tetra Gold is a legitimate channel for the acquisition, storage, and transfer of gold assets through blockchain. Its platform is intuitive and caters to investors of several backgrounds; from novices to experienced traders.

To facilitate the tokenization of gold using smart contracts, Tetra Gold allows fractional ownership and thus a ready means for any transaction. It fosters inclusion by promoting gold investment for a wider demography of investors without geographical restrictions.

Tetra Gold ensures safety, transparency, and traceability with a reliable platform that empowers people to invest in one of the most timeless and priceless commodities in existence safely.